% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allBinaryTreeShapes.R
\name{allBinaryTreeShapes}
\alias{allBinaryTreeShapes}
\title{Calculate all full unordered binary trees up to n tips}
\usage{
allBinaryTreeShapes(numTips, type = "real", progressBar = FALSE,
  startingPoint)
}
\arguments{
\item{type}{"real", "complex" or "phylo"}

\item{numtips}{max number of tips}
}
\value{
list of lists containing the trees
}
\description{
Return complex coefficient matrices, real coefficient matrices or phylo objects for all possible full binary trees up to n tips.
}
\examples{

# phylo type example

}
