% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allTrees.R
\name{allTrees}
\alias{allTrees}
\title{Calculate all coefficient matrices for trees with n tips}
\usage{
allTrees(numTips)
}
\arguments{
\item{numtips}{the max number of tips to find all coefficient matrices for}
}
\value{
list containing all possible coefficient matrices
}
\description{
Return coefficient matrices for all possible full binary trees with 2 to n tips.
}
\examples{
# example goes here
}
