% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFunctions.R
\name{coeffDist}
\alias{coeffDist}
\title{Calculates the distance matrix from multiple coefficient matrices}
\usage{
coeffDist(coefficientMatrices, method = "logL1", progressBar = FALSE,
  smallDistanceMatrix = FALSE)
}
\arguments{
\item{coefficientMatrices}{list of complex or real coefficient matrices}

\item{method}{method to use when calculating coefficient distances: "logL1", "wLogL1", "b"  or "c"}

\item{progressBar}{add a progress bar to track progress}

\item{smallDistanceMatrix}{whether to return a distance matrix (TRUE) or a distance value (FALSE) if only two coefficient matrices are given}
}
\value{
distance matrix calculated from argument coefficient matrices
}
\description{
Calculates the distance matrix from multiple coefficient matrices
}
\note{
the complex coefficient vector only supports the "logL1" method

by default a list of two coefficient matrices will return a distance value rather than a 2*2 distance matrix
}
\examples{




}
