% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficientMetrics.R
\name{coefficientAlign}
\alias{coefficientAlign}
\title{Aligns a list of coefficent matrices to the max size in the argument list}
\usage{
coefficientAlign(coeffMats)
}
\arguments{
\item{coeffMats}{list of real or complex coefficient matrices of different sizes to be aligned}
}
\value{
the aligned list of coefficient matrices
}
\description{
Aligns a list of coefficent matrices to the max size in the argument list
}
\examples{
# example goes here
}
