% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficientMetrics.R
\name{coefficientDist}
\alias{coefficientDist}
\title{Calculates the distance matrix of multiple coefficient matrices}
\usage{
coefficientDist(coefficientMatrices, method = "canberra")
}
\arguments{
\item{coefficientMatrices}{list of coefficient matrices}

\item{method}{method to use when calculating coefficient distances}
}
\value{
distance matrix calculated from argument coefficient matrices
}
\description{
Uses a specified method to construct a distance matrix from coefficient matrices.
For a list of methods see \link[parallelDist]{parDist}.
}
\examples{
# example goes here
}
