% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treenomial.R
\name{coefficientMatrix}
\alias{coefficientMatrix}
\title{Converts trees (objects of class "phylo") to coefficient matrices}
\usage{
coefficientMatrix(tree, complex = FALSE, tipLabel = FALSE,
  normalize = FALSE, progressBar = FALSE, cl = NULL)
}
\arguments{
\item{tree}{an object or list of class "phylo"}

\item{complex}{whether or not to return the polynomial with y = i}

\item{normalize}{whether or not to use a row sum normilization of the coefficients}

\item{progressBar}{add a progress bar to track performance}
}
\value{
The resulting coefficient matrix or matrices.
}
\description{
Converts trees (objects of class "phylo") to coefficient matrices
}
\examples{
library(ape)
# generate a tree
tree <- rtree(n = 500, rooted = TRUE)
# get the coefficient matrix describing the polynomial for this tree
coefficientMatrix(tree)
}
