% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffMatrix.R
\name{coeffMatrix}
\alias{coeffMatrix}
\title{Converts trees (objects of class "phylo") to real coefficient matrices or complex coefficient vectors}
\usage{
coeffMatrix(trees, type = "real", progressBar = FALSE, cl = NULL)
}
\arguments{
\item{trees}{an object or list of class "phylo"}

\item{type}{"real" to return a coefficient matrix of the real polynomial or "complex" for the complex coefficietn vector of the real polynomial with y = 1 + i}

\item{progressBar}{add a progress bar}

\item{cl}{A cluster object created by makeCluster, from the parallel package}
}
\value{
The resulting coefficient matrix or matrices.
}
\description{
Converts trees (objects of class "phylo") to real coefficient matrices or complex coefficient vectors
}
\examples{
require(ape)
# generate a tree
tree <- rtree(n = 30, rooted = TRUE)
# a real coefficient matrix
coefficientMatrix(tree)

# complex coefficient vector for the tree
coefficientMatrix(tree, complex = TRUE)

# for a list of trees with a progress bar
coeffMatrix(rmtree(10, 30), progressBar = TRUE)

# use the parallel package to speed up conversion for a large list of phylo objects
require(parallel)
cl <- makeCluster(detectCores())
largeForest <- coeffMatrix(rmtree(400, 500), progressBar = TRUE, cl = cl)
stopCluster(cl)
}
