% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterEstimation.R
\name{estimateP}
\alias{estimateP}
\title{Estimates p}
\usage{
estimateP(tree, start = 1.5, end = 8.5, step = 1, numStepSim = 100,
  numTotalSim = 10)
}
\arguments{
\item{tree}{list of coefficient matrices}
}
\value{
estimated p
# example goes here
}
\description{
Estimates p
}
