% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterEstimation.R
\name{estimateR0}
\alias{estimateR0}
\title{Estimates R0}
\usage{
estimateR0(tree, start = 1.5, end = 8.5, step = 1, numSim = 100)
}
\arguments{
\item{tree}{list of coefficient matrices}
}
\value{
estimated R0
# example goes here
}
\description{
Estimates R0
}
