% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice.R
\name{lattDist}
\alias{lattDist}
\title{Calculates the distance between two tree lattices}
\usage{
lattDist(L1, L2, w = 1/4)
}
\arguments{
\item{L1, L2}{lattices to calculate distance between}

\item{w}{the amount to weight absent positions in each lattice in the distance}
}
\value{
distance value
}
\description{
Calculates the distance between two tree lattices
}
\examples{

library(treenomial)
library(ape)


L1 <- treeToLattice(rtree(10), numThreads = 0)
L2 <- treeToLattice(rtree(10), numThreads = 0)
d <- lattDist(L1, L2, w = 1/4)


}
