% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice.R
\name{lattToDistMat}
\alias{lattToDistMat}
\title{Calculates the distance matrix from a list of tree lattices}
\usage{
lattToDistMat(lattices, w = 1/4, numThreads = -1)
}
\arguments{
\item{lattices}{list of tree lattices}

\item{w}{the amount to weight absent positions in each lattice in the distance}

\item{numThreads}{number of threads to be used, the default (-1) will use the number of cores in the machine and numThreads = 0 will only use the main thread}
}
\value{
distance matrix calculated from argument tree lattices
}
\description{
Calculates the distance matrix from a list of tree lattices
}
\examples{

library(treenomial)
library(ape)

# lattices for ten trees of 10 tips
latts <- treeToLattice(rmtree(10, 10), numThreads = 0)

# distance matrix from the list of lattices
d <- lattToDistMat(latts, numThreads = 0)

}
