% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFunctions.R
\name{phyloDistPlot}
\alias{phyloDistPlot}
\title{Plot the min/max distance tree from a target tree in a distance matrix}
\usage{
phyloDistPlot(trees, distMatrix, target, comparison = "min",
  plotFacing = FALSE, returnNearestInfo = FALSE)
}
\arguments{
\item{trees}{list of phylo objects}

\item{comparison}{find the "min" or the "max" distance tree in the distance matrix}

\item{plotFacing}{whether to plot the trees with the tips facing each other}

\item{returnNearestInfo}{whether to return the smallest/max distance phylo object and its distance to target (TRUE) or have the function return no value (FALSE)}
}
\description{
Plot the min/max distance tree from a target tree in a distance matrix
}
\examples{

}
