% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFunctions.R
\name{treeDist}
\alias{treeDist}
\title{Calculates the distance between coefficient matrices}
\usage{
treeDist(x, Y, type = "real", method = "logDiff")
}
\arguments{
\item{x}{single coefficient matrix to find distances to}

\item{Y}{list or single coefficient matrix}

\item{type}{one of:
\describe{
  \item{\dQuote{real}}{real coefficient polynomials}
  \item{\dQuote{complex}}{complex coefficient polynomials (real polynomials with y = 1 + i)}
  \item{\dQuote{tipLabel}}{complex coefficient polynomial that utilize binary trait tip labels on the phylo objects}
}}

\item{method}{method to use when calculating coefficient distances:
\describe{
  \item{\dQuote{logDiff}}{for two coefficient matrices A and B returns sum(log(1+abs(A-B))}
  \item{\dQuote{wLogDiff}}{performs the \dQuote{logDiff} method with weights on the rows}
  \item{\dQuote{pa}}{total pairs where the coefficient is present in one matrix and absent in the other (presence-absence)}
  \item{\dQuote{ap}}{opposite comparison of pa (absence-presence)}
}}
}
\value{
vector of distances
}
\description{
Calculates the distance between two coefficient matrices or a coefficient matrix and a list of coefficient matrices.
}
\note{
the complex coefficient vector and the complex tip label coefficient matrix only supports the \dQuote{logDiff} method
}
\examples{


}
