% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficientMetrics.R
\name{treenomialMDS}
\alias{treenomialMDS}
\title{Builds a 2d or 3d mds plot from tree or coefficient matrix data}
\usage{
treenomialMDS(trees, coefficientMatrices, method = "canberra", dim = 2)
}
\arguments{
\item{trees}{data frame of phylo objects used to calculate coefficient matrices}

\item{coefficientMatrices}{list of coefficient matrices}

\item{method}{method to use when caculating coefficient distances}

\item{dim}{number of dims of resulting mds plot}
}
\value{
2d or 3d MDS plot
}
\description{
Either calculates or recieves coefficient matrix data of phylo objects and uses a distance method
on the coefficients to construct an mds plot.
}
\examples{
# example goes here
}
