% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice.R
\name{treeToLattice}
\alias{treeToLattice}
\title{Convert trees to lattice form}
\usage{
treeToLattice(trees, numThreads = -1)
}
\arguments{
\item{trees}{a single phylo object or a list of phylo objects}

\item{numThreads}{number of threads to be used, the default (-1) will use the number of cores in the machine and numThreads = 0 will only use the main thread}
}
\value{
the resulting lattice or lattices as matrices with columns:
\describe{
  \item{\dQuote{node}}{the node number}
  \item{\dQuote{pl}}{the parent of this node}
  \item{\dQuote{lattice}}{the position of this node in the tree lattice}
  \item{\dQuote{bl}}{branch length to this node from its parent}
  \item{\dQuote{depth}}{depth of this node in the tree}
}
}
\description{
Converts rooted full binary trees to a tree lattice using tree distinguishing polynomials.
}
\examples{
library(treenomial)
library(ape)

tree <- rtree(10)
lattice <- treeToLattice(tree, numThreads = 0)

}
