% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedge.R
\name{wedge}
\alias{wedge}
\title{Performs the wedge operation on two coefficient matrices}
\usage{
wedge(coefficientMatrixA, coefficientMatrixB)
}
\arguments{
\item{coefficientMatrixA}{First tree coefficient matrix}

\item{coefficientMatrixB}{Second tree coefficient matrix}
}
\value{
The resulting polynomail matrix.
}
\description{
Calculates the result from the wedge operation of two coefficient matrices.
}
\examples{
leaf <- sparseMatrix(1, 2, x = 1)
wedge(leaf, leaf)
[1,] . . 1
[2,] 1 . .
}
