% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeToPoly.R
\name{alignPoly}
\alias{alignPoly}
\title{Align varios types of coefficient matrices}
\usage{
alignPoly(coefficientMatrices)
}
\arguments{
\item{coefficientMatrices}{a list of coefficient matrices of various sizes}
}
\value{
the aligned list of coefficient matrices
}
\description{
Align varios types of coefficient matrices
}
\examples{
library(treenomial)
library(ape)
differentSizeTrees <- c(rtree(2), rmtree(10,10))
coeffs <- treeToPoly(differentSizeTrees)
alignedCoeffs <- alignPoly(coeffs)

}
