% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allBinaryTreeShapes.R
\name{allBinaryTreeShapes}
\alias{allBinaryTreeShapes}
\title{Calculate all full unordered binary trees up to n tips}
\usage{
allBinaryTreeShapes(numTips, type = "real", progressBar = FALSE,
  startingPoint)
}
\arguments{
\item{type}{"real", "complex" or "phylo"}

\item{progressBar}{add a progress bar}

\item{startingPoint}{provide a previous list of all the possible unordered full binary trees to start the calculation from}

\item{numtips}{max number of tips}
}
\value{
list of lists containing the trees
}
\description{
Return complex coefficient matrices, real coefficient matrices or phylo objects for all possible unordered full binary trees up to n tips.
}
\examples{

# generate coefficient matrices describing the polynomials of all possible unordered full binary trees
# up to 10 tips and print out the number of matrices at each tip number.
allBinTenRealCoeff <- allBinaryTreeShapes(10, type = "phylo")
lengths(allBinTenRealCoeff)

# phylo type example, plot all 6 tip unordered full binary trees
allBinSixPhylo <- allBinaryTreeShapes(6, type = "phylo")[[6]]
par(mfrow=c(1,6))
lapply(allBinSixPhylo, function(t) plot.phylo(ladderize(t), direction = "downwards", show.tip.label = FALSE))

}
