% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treenomial.R
\name{coefficientMatrix}
\alias{coefficientMatrix}
\title{Converts trees (objects of class "phylo") to coefficient matrices}
\usage{
coefficientMatrix(tree, complexMode = FALSE)
}
\arguments{
\item{tree}{an object or list of class "phylo"}
}
\value{
The resulting coefficient matrix or matrices.
}
\description{
Converts trees (objects of class "phylo") to coefficient matrices
}
\examples{
# example goes here
}
