% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficientMatrix.R
\name{coefficientMatrix}
\alias{coefficientMatrix}
\title{Converts trees (objects of class "phylo") to real or complex coefficient matrices}
\usage{
coefficientMatrix(trees, type = "real", progressBar = FALSE,
  cl = NULL)
}
\arguments{
\item{trees}{an object or list of class "phylo"}

\item{type}{"real" for real polynomial or "complex" for polynomial with y = 1 + i}

\item{progressBar}{add a progress bar to track performance}
}
\value{
The resulting coefficient matrix or matrices.
}
\description{
Converts trees (objects of class "phylo") to real or complex coefficient matrices
}
\examples{
library(ape)
# generate a tree
tree <- rtree(n = 30, rooted = TRUE)
# get the coefficient matrix describing the real coefficient polynomial for this tree
coefficientMatrix(tree)

# generate the complex coefficient matrix describing the complex coefficient polynomial for the tree
coefficientMatrix(tree, complex = TRUE)

# for a list of trees



}
