% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFunctions.R
\name{phyloDist}
\alias{phyloDist}
\title{Calculates the distance matrix from a list of trees}
\usage{
phyloDist(trees, method = "logL1", type = "real",
  progressBar = FALSE, smallDistanceMatrix = FALSE, cl = NULL)
}
\arguments{
\item{trees}{list of phylo objects}

\item{method}{method to use when calculating coefficient distances: "logL1", "wLogL1", "b"  or "c"}

\item{progressBar}{add a progress bar to track progress}

\item{smallDistanceMatrix}{whether to return a distance matrix (TRUE) or a distance value (FALSE) if only two coefficient matrices are given}

\item{cl}{a cluster object created by makeCluster}
}
\value{
distance matrix calculated from argument coefficient matrices
}
\description{
Calculates the distance matrix from a list of trees
}
\note{
the complex coefficient vector only support the "logL1" method

by default a list of two coefficient matrices will return a distance value rather than a 2*2 distance matrix
}
