% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficientMetrics.R
\name{treenomialMDS}
\alias{treenomialMDS}
\title{Builds a 2d or 3d mds plot from tree or coefficient matrix data}
\usage{
treenomialMDS(trees, coefficientMatrices, method = "canberra", dim = 2)
}
\arguments{
\item{trees}{data frame of phylo objects used to calculate coefficient matrices}

\item{coefficientMatrices}{list of coefficient matrices}

\item{method}{method to use when caculating coefficient distances}

\item{dim}{number of dims of resulting mds plot}
}
\value{
2d or 3d MDS plot
}
\description{
Either calculates or recieves coefficient matrix data of phylo objects and uses a distance method
on the coefficients to construct an mds plot.
}
\examples{
# generate some forests
numTrees <- 50
numTips <- 500

pdaTrees <- rtreeshape(numTrees,tip.number=numTips,model="pda")
yuleTrees <- rtreeshape(numTrees,tip.number=numTips,model="yule")
aldousTrees <- rtreeshape(numTrees,tip.number=numTips,model="aldous")
biasedTrees <- rtreeshape(numTrees,tip.number=numTips,model="biased")

# place into a list
allTrees <- list(pdaTrees = pdaTrees, yuleTrees = yuleTrees, aldousTrees = aldousTrees, biasedTrees, biasedTrees)

# convert to phylo type
allTrees <- lapply(unlist(allTrees, recursive=FALSE), as.phylo)

# construct the 2d MDS on the trees using the "canberra" distance method
mds2D <- treenomialMDS(allTrees, method = "canberra", dim = 2)

# first construct the coefficient matrices and use them to create two MDS plots
coeffMats <- coefficientMatrix(allTrees)
mds2dTwoStep <- treenomialMDS(coeffMats, method = "canberra", dim = 2)
mds3dTwoStep <- treenomialMDS(coeffMats, method = "canberra", dim = 3)

}
