% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedge.R
\name{wedge}
\alias{wedge}
\title{Performs the wedge operation}
\usage{
wedge(A, B, type = "real")
}
\arguments{
\item{A}{First coefficient matrix or phylo object}

\item{B}{Second coefficient matrix or phylo object}

\item{type}{The type of objects to wedge: "real", "complex" or "phylo"}
}
\value{
Wedge result coefficient matrix or phylo object
}
\description{
Calculates the result from the wedge operation on two real coefficient
matrices, two complex coefficient matrices or two phylo objects.
}
\examples{
require(Matrix)
require(ape)

# wedge two real coefficient matrices
leaf <- sparseMatrix(1, 2, x = 1)
wedge(leaf, leaf, type = "real")

# wedge two complex coefficient matrices
leaf <- as.complex(c(0,1))
wedge(leaf, leaf, type = "complex")

# wedge two phylo objects
leaf <- "leaf"
wedge(leaf, leaf, type = "phylo")

}
