% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedge.R
\name{wedgeComplex}
\alias{wedgeComplex}
\title{Performs the wedge operation on two complex coefficient matrices}
\usage{
wedgeComplex(op1, op2)
}
\arguments{
\item{op1}{First tree complex coefficient matrix}

\item{op2}{Second tree complex coefficient matrix}
}
\value{
The resulting coefficient matrix.
}
\description{
Calculates the result from the wedge operation of two complex coefficient matrices.
}
\examples{
example here
}
